refactor(cloudquery): Extract Cluster into it's own class
This'll make it easier to add to the existing EC2 stack.
